Almost done with hurtbox and hitbox

+++ Game/Custom/Components/Health.cs
@@ -3,10 +3,10 @@ namespace Game.Custom.Components;
 
 public class HealthComponent(int maxHealth)
 {
-    private int _health = maxHealth;
-    public int MaxHealth = maxHealth;
+    private float _health = maxHealth;
+    public float MaxHealth = maxHealth;
 
-    public int Health { get => _health; set => _health = (value > MaxHealth) ? MaxHealth : value; }
+    public float Health { get => _health; set => _health = (value > MaxHealth) ? MaxHealth : value; }
     public bool IsAlive => _health > 0;
 }
